+++ PREACT/PREACTcore/Source/TriggerBuffer/TriggerBufferModule.cs
@@ -2,9 +2,9 @@ namespace PREACT
 {
     public abstract class TriggerBufferModule
     {
-        protected int[,] _triggerBufferOutput;
+        protected float[,] _triggerBufferOutput;
 
-        public int[,] TriggerBufferOutput { get => _triggerBufferOutput; }    
+        public float[,] TriggerBufferOutput { get => _triggerBufferOutput; }    
 
 
         public abstract void Run();
